Implement the Master Server and improve the client.
These changes follows the structure of the client with a bit of a change. It also has a ServerClient which does just about the same as a Client while exposing a bit of its senders to allow the MasterServer to send messages through it. It also strips down the amount of data that's being stored. We don't need to store the ID, that's tied to the index and the loop already. It's redundant. We also don't need to store the name of the client on the master server. Maybe on the cluster. But on the Master Server we're just redirecting people.

It also adds some notes on the approach for the auth server. Finally! :)